fix: report declared configuration accurately in triage
